add squadron sql test of mp repo tests
