+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Repositories/ChargeRepository.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Application.ChangeOfCharges.Repositories;
 using GreenEnergyHub.Charges.Domain.ChangeOfCharges.Transaction;
