+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.TestCore/Squadron/SquadronContextFactory.cs
@@ -1,0 +1,70 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Threading.Tasks;
+using GreenEnergyHub.Charges.ApplyDBMigrationsApp.Helpers;
+using GreenEnergyHub.Charges.Infrastructure.Context;
+using Microsoft.EntityFrameworkCore;
+using Squadron;
+
+namespace GreenEnergyHub.Charges.TestCore.Squadron
+{
+    public static class SquadronContextFactory
+    {
+        public static async Task<ChargesDatabaseContext> GetDatabaseContextAsync(SqlServerResource<SqlServerOptions> resource)
+        {
+            if (resource is null) throw new ArgumentNullException(nameof(resource));
+
+            var connectionString = await CreateDatabaseAsync(resource)
+                    .ConfigureAwait(false);
+            DbContextOptions<ChargesDatabaseContext> dbContextOptions =
+                new DbContextOptionsBuilder<ChargesDatabaseContext>()
+                    .UseSqlServer(connectionString)
+                    .Options;
+
+            var chargesContext = new ChargesDatabaseContext(dbContextOptions);
+
+            return chargesContext;
+        }
+
+        public static async Task EmptyDatabaseAsync(ChargesDatabaseContext chargesDatabaseContext)
+        {
+            if (chargesDatabaseContext == null) throw new ArgumentNullException(nameof(chargesDatabaseContext));
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[ChargePrice]").ConfigureAwait(false);
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[ChargeOperation]").ConfigureAwait(false);
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[ChargePeriodDetails]").ConfigureAwait(false);
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[MarketParticipant]").ConfigureAwait(false);
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[Charge]").ConfigureAwait(false);
+            await chargesDatabaseContext.Database.ExecuteSqlRawAsync("DELETE FROM [Charges].[MeteringPoint]").ConfigureAwait(false);
+        }
+
+        private static async Task<string> CreateDatabaseAsync(SqlServerResource<SqlServerOptions> resource)
+        {
+            const string databaseName = "New_Database";
+
+            var connectionString = await resource.CreateDatabaseAsync(databaseName).ConfigureAwait(false);
+
+            var upgrader = UpgradeFactory.GetUpgradeEngine(connectionString, _ => true);
+            var result = upgrader.PerformUpgrade();
+
+            if (result.Successful is false)
+            {
+                throw new Exception("Database migration failed");
+            }
+
+            return connectionString;
+        }
+    }
+}
