+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Infrastructure/Repositories/MeteringPointRepositoryTests.cs
@@ -12,17 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Threading.Tasks;
 using FluentAssertions;
-using GreenEnergyHub.Charges.ApplyDBMigrationsApp.Helpers;
 using GreenEnergyHub.Charges.Domain.MeteringPoints;
-using GreenEnergyHub.Charges.Infrastructure.Context;
 using GreenEnergyHub.Charges.Infrastructure.Repositories;
 using GreenEnergyHub.Charges.TestCore.Squadron;
-using Microsoft.EntityFrameworkCore;
 using NodaTime;
 using Squadron;
 using Xunit;
@@ -44,7 +38,8 @@ namespace GreenEnergyHub.Charges.Tests.Infrastructure.Repositories
         public async Task StoreMeteringPointAsync_WhenMeteringPointIsCreated_StoresMeteringPointInDatabase()
         {
             // Arrange
-            await using var chargesDatabaseContext = await GetDatabaseContext(_resource)
+            await using var chargesDatabaseContext = await SquadronContextFactory
+                .GetDatabaseContextAsync(_resource)
                 .ConfigureAwait(false);
             {
                 var expected = GetMeteringPointCreatedEvent();
@@ -72,37 +67,6 @@ namespace GreenEnergyHub.Charges.Tests.Infrastructure.Repositories
                 SystemClock.Instance.GetCurrentInstant(),
                 ConnectionState.Connected,
                 SettlementMethod.Flex);
-        }
-
-        private static async Task<ChargesDatabaseContext> GetDatabaseContext(SqlServerResource<SqlServerOptions> resource)
-        {
-            var connectionString = await CreateDatabaseAsync(resource)
-                .ConfigureAwait(false);
-            DbContextOptions<ChargesDatabaseContext> dbContextOptions =
-                new DbContextOptionsBuilder<ChargesDatabaseContext>()
-                    .UseSqlServer(connectionString)
-                    .Options;
-
-            var chargesContext = new ChargesDatabaseContext(dbContextOptions);
-
-            return chargesContext;
-        }
-
-        private static async Task<string> CreateDatabaseAsync(SqlServerResource<SqlServerOptions> resource)
-        {
-            const string databaseName = "New_Database";
-
-            var connectionString = await resource.CreateDatabaseAsync(databaseName).ConfigureAwait(false);
-
-            var upgrader = UpgradeFactory.GetUpgradeEngine(connectionString, _ => true);
-            var result = upgrader.PerformUpgrade();
-
-            if (result.Successful is false)
-            {
-                throw new Exception("Database migration failed");
-            }
-
-            return connectionString;
         }
     }
 }
